pyproject.toml
[project]
name = "marketwise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deep-sort-realtime>=1.3.2",
    "deepface>=0.0.97",
    "opencv-python>=4.12.0.88",
    "psycopg2-binary>=2.9.11",
    "setuptools>=80.9.0",
    "ultralytics>=8.3.253",
]
